+++ algorithm/src/main/java/dev/carbonshow/algorithm/partition/IntegerPartition.java
@@ -11,18 +11,18 @@ public interface IntegerPartition {
   /**
    * 获取最大划分方案数量
    *
-   * @param addendSet 加数集合，内部不能有重复的元素，必须均为正整数，同时应该按升序排列。不同数量的不同加数之和应该等于 `partitioned`
+   * @param addendSet   加数集合，内部不能有重复的元素，必须均为正整数，同时应该按升序排列。不同数量的不同加数之和应该等于 `partitioned`
    * @param partitioned 被划分的正整数
    * @return 返回划分方案的最大数量
    */
-  int solve(int[] addendSet, int partitioned);
+  long solve(long[] addendSet, long partitioned);
 
   /**
    * 获取最大划分方案的细节，即具体划分方法
    *
-   * @param addendSet 加数集合，内部不能有重复的元素，必须均为正整数。不同数量的不同加数之和应该等于 `partitioned`
+   * @param addendSet   加数集合，内部不能有重复的元素，必须均为正整数。不同数量的不同加数之和应该等于 `partitioned`
    * @param partitioned 被划分的正整数
    * @return 返回所有的划分方案，一维区分不同划分方案，二维表示一个划分方案中的加数选择。该方案集合一定对应最大划分数。
    */
-  ArrayList<ArrayList<Integer>> solveWithPartitions(int[] addendSet, int partitioned);
+  ArrayList<ArrayList<Long>> solveWithPartitions(long[] addendSet, long partitioned);
 }
